LayoutTests/js/typedarray-reduceRight-expected.txt

This test checks the behavior of the TypedArray.prototype.reduceRight function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.reduceRight.length is 1
PASS Int32Array.prototype.reduceRight.name is 'reduceRight'
PASS isSameFunctionForEachTypedArrayPrototype('reduceRight') is true
PASS testPrototypeReceivesArray('reduceRight', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

1.0 Single Argument Testing
PASS testPrototypeFunction('reduceRight', '(createArray)', [12, 5, 8, 13, 44], [12, 5, 8, 13, 44].reverse(), [12, 5, 8, 13, 44]) is true
PASS testPrototypeFunction('reduceRight', '(sum)', [1, 2, 3, 4, 5], 15) is true

2.0 Two Argument Testing
PASS testPrototypeFunction('reduceRight', '(createArray, [1])', [12, 23, 11, 1, 45], [1, 45, 1, 11, 23, 12], [12, 23, 11, 1, 45]) is true

3.0 Array Element Changing
PASS testPrototypeFunction('reduceRight', '(createArrayAndChange, [])', [12, 15, 2, 13, 44], [44, 13, 2, 5, 5], [5, 5, 5, 5, 5]) is true

4.0 Exception Test
PASS testPrototypeFunction('reduceRight', '(isBigEnoughAndException)', [12, 15, 10, 13, 44], false) threw exception exception from function.

5.0 Wrong Type for Callback Test
PASS testPrototypeFunction('reduceRight', '(8)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduceRight callback must be a function.
PASS testPrototypeFunction('reduceRight', '("wrong")', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduceRight callback must be a function.
PASS testPrototypeFunction('reduceRight', '(new Object())', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduceRight callback must be a function.
PASS testPrototypeFunction('reduceRight', '(null)', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduceRight callback must be a function.
PASS testPrototypeFunction('reduceRight', '()', [12, 15, 10, 13, 44], false) threw exception TypeError: TypedArray.prototype.reduceRight callback must be a function.

PASS successfullyParsed is true

TEST COMPLETE
